Now inject monkeytype.js bundle only in development mode
