+++ frontend/webpack/config.dev.js
@@ -1,5 +1,22 @@
+const { resolve } = require("path");
 const { merge } = require("webpack-merge");
 const BASE_CONFIG = require("./config.base");
+const HtmlWebpackPlugin = require("html-webpack-plugin");
+
+const htmlWebpackPlugins = [
+  "terms-of-service",
+  "security-policy",
+  "privacy-policy",
+  "email-handler",
+  "das",
+].map((name) => {
+  return new HtmlWebpackPlugin({
+    filename: `${name}.html`,
+    template: resolve(__dirname, `../static/${name}.html`),
+    inject: "body",
+    cache: false,
+  });
+});
 
 /** @type { import('webpack').Configuration } */
 const DEV_CONFIG = {
@@ -16,6 +33,8 @@ const DEV_CONFIG = {
       overlay: false,
     },
   },
+
+  plugins: htmlWebpackPlugins,
 };
 
 module.exports = merge(BASE_CONFIG, DEV_CONFIG);
